test: move DirInfo.abs_path test to its own file with isolated cache
The sibling run-autoinstall.test.ts tests mutate the shared bunEnv and
routinely time out at the 5s default on debug/ASAN builds when hitting
the live registry, which masked this test's pass/fail signal during
gate validation.

Use a per-test BUN_INSTALL_CACHE_DIR: a shared cache can be left
half-populated (extracted dir present, name/version symlink missing)
if a prior run is killed mid-install, which makes pathForCachedNPMPath's
readlinkat return ENOENT before auto-install kicks in. Add a 30s timeout
for the cold-cache download on debug builds, matching bun-add.test.ts /
bun-create.test.ts.
